Rename stream base worker
