+++ ec-2/nexusflow/lib/queue.ts
@@ -1,4 +1,4 @@
 import { FlowProducer } from 'bullmq';
-import { redisClient } from '../workers/streamBase';
+import { redisClient } from '../workers/streamAwareBullWorker';
 
 export const flowProducer = new FlowProducer({ connection: redisClient });
